1/2、强生 出租_ERP系统 单表添加 accident_metting_safe 表失败bug修复，原因：上游库中，字段长度大于代码使用的int类型

+++ fisk-consume-serveice/fisk-consume-serveice-web/src/main/java/com/fisk/dataservice/util/SqlServerPlusUtils.java
@@ -289,7 +289,7 @@ public class SqlServerPlusUtils {
                 // 字段类型
                 tableStructureDTO.fieldType = metaData.getColumnTypeName(i);
                 // 字段长度
-                tableStructureDTO.fieldLength = metaData.getColumnDisplaySize(i);
+                tableStructureDTO.fieldLength = (long) metaData.getColumnDisplaySize(i);
                 colNameList.add(tableStructureDTO);
             }
         } catch (SQLException e) {
